sentiment_analysis/requirements.txt
audeer==1.8.0
audformat==0.11.6
audinterface==0.7.0
audiofile==0.4.1
audobject==0.5.1
audresample==0.1.6
google-api-core==1.31.0
google-cloud-speech==2.5.0
iso-639==0.4.5
libcst==0.3.19
mypy-extensions==0.4.3
opensmile==2.2.0
oyaml==1.0
pandas==1.2.5
pyyaml==5.4.1
six==1.16.0
sox==1.4.1
threaded==4.0.8
tqdm==4.54.1
typing-inspect==0.7.1
aiohttp==3.6.3
aniso8601==8.0.0
appdirs==1.4.4
async-timeout==3.0.1
attrs==20.2.0
audioread==2.1.9
blas
brotlipy==0.7.0
bzip2
c-ares
ca-certificates
cachetools==4.1.1
certifi==2021.5.30
cffi==1.14.3
chardet==3.0.4
click==7.1.2
cryptography==3.2.1
cudatoolkit
cycler==0.10.0
dbus
decorator==4.4.2
expat
ffmpeg
flask==1.1.1
flask-restful==0.3.8
flask-sqlalchemy==2.4.0
fontconfig
freetype
gettext
glib
gmp
gnutls
google-auth==1.33.0
google-cloud-core==1.7.1
google-cloud-translate==3.2.0
googleapis-common-protos==1.53.0
grpcio==1.31.0
gst-plugins-base
gstreamer
gunicorn==19.9.0
icu==58.2
idna==2.10
intel-openmp
itsdangerous==1.1.0
jinja2==2.11.2
joblib==0.17.0
jpeg==9b
kiwisolver==1.2.0
lame==3.100
lcms2==2.12
ld_impl_linux-64==2.33.1
libcxx==11.0.0
libcxxabi==11.0.0
libedit==3.1.20191231
libffi==3.3
libflac==1.3.3
libgcc-ng==11.2.0
libgfortran-ng==7.3.0
libgomp==11.2.0
libiconv==1.15
libllvm10==10.0.1
libogg==1.3.4
libopus==1.3.1
libpng==1.6.37
libprotobuf==3.14.0
librosa==0.8.0
libsndfile==1.0.30
libstdcxx-ng==11.2.0
libtiff==4.2.0
libuuid==1.0.3
libuv==1.40.0
libvorbis==1.3.7
libwebp-base==1.2.0
libxcb==1.14
libxml2==2.9.10
llvm-openmp==10.0.0
llvmlite==0.34.0
lz4-c==1.9.3
markupsafe==1.1.1
matplotlib==3.2.2
matplotlib-base==3.2.2
mkl==2019.4
mkl-service==2.3.0
mkl_fft==1.2.0
mkl_random==1.1.0
multidict==4.7.6
ncurses==6.2
nettle==3.4.1
ninja==1.10.1
numba==0.51.2
numpy==1.19.2
numpy-base==1.19.2
olefile==0.46
openh264==1.8.0
openjpeg==2.3.0
openssl==1.1.1l
packaging==20.4
paho-mqtt==1.5.1
pcre==8.44
pillow==8.3.1
pip==20.2.4
pooch==1.6.0
proto-plus==1.19.6
protobuf==3.14.0
pyasn1==0.4.8
pyasn1-modules
pycparser==2.20
pycrypto==2.6.1
pydub==0.25.1
pyopenssl==20.0.1
pyparsing==2.4.7
pyqt==5.9.2
pysocks==1.7.1
pysoundfile==0.10.3.post1
python==3.7.9
python-dateutil==2.8.1
python_abi==3.7
pytorch==1.7.0
pytz==2020.1
pyu2f==0.1.5
qt==5.9.7
readline==8.0
requests==2.24.0
resampy==0.2.2
rsa==4.7.2
scikit-learn==0.23.2
scipy==1.5.2
setuptools==50.3.1
sip==4.19.24
sqlalchemy==1.3.8
sqlite==3.33.0
threadpoolctl==2.1.0
tk==8.6.10
torchaudio==0.7.0
torchtext==0.8.0
torchvision==0.8.0
tornado==6.0.4
typing_extensions==3.7.4.3
urllib3==1.25.11
werkzeug==1.0.1
wheel==0.35.1
x264
xlrd==2.0.1
xz==5.2.5
yarl==1.6.2
zlib==1.2.11
zstd==1.4.9

